feat: give news feeds a cover instead of an empty box
Reported: the News grid draws an empty cell with a small book icon for
every feed.

Nothing failed. News entries carry no cover art -- there is no image link
in the feed at all -- so there was never anything to download and nothing
would have appeared later. What was wrong is that "no cover" and "cover
failed to load" looked identical, and the generic BookIcon reads as the
second: an unopened book that did not work, rather than a thing you are
meant to press.

My Books solved this already for the synthetic app tiles, whose paths are
never real files: drawTileCover paints the label on a framed black cell
so it reads as a designed cover. That helper was private to
RecentBooksActivity; it is now src/components/TileCover.{h,cpp} and the
OPDS grid uses it for news feeds.

It gained a truncation step on the way. Shrinking the font twice is
enough for "Games" and not for "Ars Technica - All content", which ran
past the frame; the label is now clipped to fit.

Only the news root is affected. A catalogue book with no cover keeps the
book icon, since there the art usually does exist and may simply not have
downloaded yet -- which is a different thing to say.

Verified in the simulator against a local feed: three tiles, Arabic
shaped correctly, and the long name clipped to "Ars Technica ...".
